Nex 243 frontend

+++ MSP App/API/Controllers/TenantController.cs
@@ -1,9 +1,10 @@
-﻿using API.Services;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using SharedComponents.DbServices;
-using SharedComponents.Entities;
-using SharedComponents.WebEntities.Requests.TenantRequests;
+﻿using Microsoft.AspNetCore.Mvc;
+using SharedComponents.Entities.DbEntities;
+using SharedComponents.Entities.WebEntities.Requests.TenantRequests;
+using SharedComponents.Handlers.Attributes.HasPermission;
+using SharedComponents.Handlers.Results;
+using SharedComponents.Services.APIServices.Interfaces;
+using SharedComponents.Services.DbServices.Interfaces;
 
 namespace API.Controllers
 {
@@ -12,14 +13,17 @@ namespace API.Controllers
     [ApiExplorerSettings(GroupName = "v1-Web")]
     public class TenantController : ControllerBase
     {
-        private readonly DbTenantService _dbTenantService;
+        private readonly IDbTenantService _dbTenantService;
+        private readonly IAPIAuthService _authService;
 
-        public TenantController(DbTenantService dbTenantService)
+        public TenantController(IDbTenantService dbTenantService, IAPIAuthService authService)
         {
             _dbTenantService = dbTenantService;
+            _authService = authService;
         }
 
         [HttpPost("")]
+        [HasPermission("Tenant.Create.Permission", PermissionType.System)]
         public async Task<IActionResult> CreateAsync([FromBody] TenantCreateRequest request)
         {
             if (ModelState.IsValid)
@@ -50,16 +54,23 @@ namespace API.Controllers
         }
 
         [HttpPut("")]
+        [HasPermission("Tenant.Update.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> UpdateAsync([FromBody] TenantUpdateRequest request)
         {
             if (ModelState.IsValid)
             {
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<TenantController>(Request.Headers["Authorization"].ToString(), request.Id))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 Tenant? tenant = await _dbTenantService.GetAsync(request.Id);
                 if (tenant == null)
                 {
                     return NotFound("Tenant not found.");
                 }
-                if(tenant.TenantInfo == null)
+                if (tenant.TenantInfo == null)
                 {
                     return NotFound("TenantInfo not found.");
                 }
@@ -83,10 +94,22 @@ namespace API.Controllers
         }
 
         [HttpDelete("{id}")]
+        [HasPermission("Tenant.Delete.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> DeleteAsync(int id)
         {
             if (ModelState.IsValid)
             {
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<TenantController>(Request.Headers["Authorization"].ToString(), id))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
+                Tenant? tenant = await _dbTenantService.GetAsync(id);
+                if (tenant == null)
+                {
+                    return NotFound("Tenant not found.");
+                }
                 if (await _dbTenantService.DeleteAsync(id))
                 {
                     return Ok($"Tenant deleted successfully.");
@@ -98,10 +121,17 @@ namespace API.Controllers
         }
 
         [HttpGet("{id}")]
+        [HasPermission("Tenant.Get.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> GetAsync(int id)
         {
             if (ModelState.IsValid)
             {
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<TenantController>(Request.Headers["Authorization"].ToString(), id))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 Tenant? tenant = await _dbTenantService.GetAsync(id);
                 if (tenant != null)
                 {
@@ -113,10 +143,17 @@ namespace API.Controllers
         }
 
         [HttpGet("{id}/Rich")]
+        [HasPermission("Tenant.Get-Rich.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> GetRichAsync(int id)
         {
             if (ModelState.IsValid)
             {
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<TenantController>(Request.Headers["Authorization"].ToString(), id))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 Tenant? tenant = await _dbTenantService.GetRichAsync(id);
                 if (tenant != null)
                 {
@@ -128,6 +165,7 @@ namespace API.Controllers
         }
 
         [HttpGet("")]
+        [HasPermission("Tenant.Get-All.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> GetAllAsync()
         {
             if (ModelState.IsValid)
